first round of search route endpoint going with specs
given the location of user and a radius to search in, this returns an array of routes (sorted by distance) around the user. For each route, it also includes the next 10 runs and their times.

Time for each run/location combo is currently a naive implementation using a string "4:30p, 4:45p, 5:00p" which should do for now but will have to be revisited in future.